requirements.txt
requests
flask
cachetools
Pillow
pytest
vcrpy
